requirements.txt
absl-py==0.14.1
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
catboost==0.25.1
category-encoders==2.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
dask==2021.4.1
debugpy==1.4.3
decorator==4.4.2
defusedxml==0.7.1
distributed==2021.4.1
entrypoints==0.3
evalml==0.23.0
fastapi==0.68.1
featuretools==0.23.3
flatbuffers==1.12
fsspec==2021.4.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.16
grpcio==1.41.0
h11==0.12.0
h5py==3.1.0
HeapDict==1.0.1
idna==3.2
imbalanced-learn==0.8.0
importlib-metadata==4.8.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lightgbm==3.0.0
llvmlite==0.36.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nlp-primitives==1.1.0
nltk==3.6.2
notebook==6.4.4
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.3.2
plotly==4.14.3
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==5.8.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.4.4
requests==2.26.0
requests-oauthlib==1.3.0
requirements-parser==0.2.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.2
scikit-optimize==0.8.1
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.8.0
shap==0.39.0
six==1.15.0
sktime==0.6.0
slicer==0.0.7
sortedcontainers==2.3.0
starlette==0.14.2
statsmodels==0.12.2
tblib==1.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-serving-api==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
texttable==1.6.3
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
woodwork==0.0.11
wrapt==1.12.1
xgboost==1.2.1
zict==2.0.0
zipp==3.6.0
